requirements.txt
tensorflow
flask
pillow
streamlit